OS-5696 lint error in contract_process_fork() OS-5697 lxbrand references readlinkat() with wrong type Reviewed by: Patrick Mooney <patrick.mooney@joyent.com> Reviewed by: Jerry Jelinek <jerry.jelinek@joyent.com> Approved by: Jerry Jelinek <jerry.jelinek@joyent.com>

+++ usr/src/uts/common/brand/lx/syscall/lx_link.c
@@ -30,7 +30,7 @@ extern int unlinkat(int, char *, int);
 /* From uts/common/syscall/symlink.c */
 extern int symlinkat(char *, int, char *);
 /* From uts/common/syscall/readlink.c */
-extern int readlinkat(int, char *, char *, size_t);
+extern ssize_t readlinkat(int, char *, char *, size_t);
 
 static long
 lx_link_common(int ffd, char *from, int tfd, char *to, int flags)
